+++ webui/src/models/cluster/failover/index.ts
@@ -7,13 +7,14 @@ import type {
 } from 'src/generated/graphql-typing-ts';
 import { app } from 'src/models';
 
-import type { Failover } from './types';
+import type { Failover, StateProviderStatus } from './types';
 
 const { some } = app.utils;
 
 // events
 export const failoverModalOpenEvent = app.domain.createEvent('failover modal open event');
 export const failoverModalCloseEvent = app.domain.createEvent('failover modal close event');
+export const stateProviderStatusGetEvent = app.domain.createEvent('state provider status popover open event');
 
 export const changeFailoverEvent = app.domain.createEvent<ChangeFailoverMutationVariables>('change failover event');
 
@@ -21,12 +22,17 @@ export const changeFailoverEvent = app.domain.createEvent<ChangeFailoverMutation
 export const $failoverModalVisible = app.domain.createStore(false);
 export const $failoverModalError = app.domain.createStore<string>('');
 export const $failover = app.domain.createStore<Failover>(null);
+export const $stateProviderStatus = app.domain.createStore<StateProviderStatus[]>([]);
 
 // effects
 export const getFailoverFx = app.domain.createEffect<void, GetFailoverParamsQuery>('get failover');
 export const changeFailoverFx = app.domain.createEffect<ChangeFailoverMutationVariables, ChangeFailoverMutation>(
   'change failover'
 );
+export const getStateProviderStatusFx = app.domain.createEffect<
+  void,
+  { cluster: { failover_state_provider_status: StateProviderStatus[] } }
+>('get state provider status');
 
 // computed
 export const $failoverModal = combine({