+++ webui/src/pages/Cluster/components/FailoverModal/components/FailoverModalForm/FailoverModalForm.tsx
@@ -1,5 +1,6 @@
 /* eslint-disable @typescript-eslint/ban-ts-comment */
 import React, { useCallback } from 'react';
+import { useEvent } from 'effector-react';
 // @ts-ignore
 import { Modal } from '@tarantool.io/ui-kit';
 
@@ -10,7 +11,8 @@ import FailoverModalFormForm, { FailoverFormProps } from './FailoverModalForm.fo
 const { failoverModalCloseEvent } = cluster.failover;
 
 const FailoverModalForm = ({ mode, failover }: FailoverFormProps) => {
-  const handleClose = useCallback(() => failoverModalCloseEvent(), []);
+  const event = useEvent(failoverModalCloseEvent);
+  const handleClose = useCallback(() => event(), [event]);
 
   return (
     <Modal visible className="meta-test__FailoverModal" title="Failover control" onClose={handleClose}>